refactor(relaxtime): extract gap transport phase equilibrium shell
